調整：將手機與伺服器程式整理至 mobile 與 server 資料夾
