+++ server/app/routes/fall_detection_routes.py
